guacamole: move terminal to protocol ssh
