+++ guacamole/Makefile.am
@@ -23,7 +23,6 @@ ACLOCAL_AMFLAGS = -I m4
 DIST_SUBDIRS =           \
     src/libguac          \
     src/common           \
-    src/terminal         \
     src/protocols/rdp    \
     src/protocols/ssh    \
     src/protocols/vnc
@@ -31,10 +30,6 @@ DIST_SUBDIRS =           \
 SUBDIRS =        \
     src/libguac  \
     src/common
-
-if ENABLE_TERMINAL
-SUBDIRS += src/terminal
-endif
 
 if ENABLE_RDP
 SUBDIRS += src/protocols/rdp
