+++ guacamole/src/protocols/ssh/terminal_char_mappings.c
