+++ guacamole/src/protocols/ssh/terminal_handlers.h
