+++ guacamole/src/protocols/ssh/terminal_named-colors.h
@@ -17,38 +17,29 @@
  * under the License.
  */
 
-
-#ifndef _GUAC_TERMINAL_COMMON_H
-#define _GUAC_TERMINAL_COMMON_H
+#ifndef GUAC_TERMINAL_NAMED_COLORS_H
+#define GUAC_TERMINAL_NAMED_COLORS_H
 
 #include "config.h"
-#include "types.h"
-
-#include <stdbool.h>
+#include "terminal_palette.h"
 
 /**
- * Returns the closest value to the value given that is also
- * within the given range.
+ * Searches for the color having the given name, storing that color within the
+ * given guac_terminal_color structure if found. If the color cannot be found,
+ * the guac_terminal_color structure is not touched. All color names supported
+ * by xterm are recognized by this function.
+ *
+ * @param name
+ *     The name of the color to search for.
+ *
+ * @param color
+ *     A pointer to the guac_terminal_color structure in which the found color
+ *     should be stored.
+ *
+ * @returns
+ *     Zero if the color was successfully found, non-zero otherwise.
  */
-int guac_terminal_fit_to_range(int value, int min, int max);
-
-/**
- * Encodes the given codepoint as UTF-8, storing the result within the
- * provided buffer, and returning the number of bytes stored.
- */
-int guac_terminal_encode_utf8(int codepoint, char* utf8);
-
-/**
- * Returns whether a codepoint has a corresponding glyph, or is rendered
- * as a blank space.
- */
-bool guac_terminal_has_glyph(int codepoint);
-
-/**
- * Similar to write, but automatically retries the write operation until
- * an error occurs.
- */
-int guac_terminal_write_all(int fd, const char* buffer, int size);
+int guac_terminal_find_color(const char* name, guac_terminal_color* color);
 
 #endif
