+++ guacamole/src/protocols/ssh/terminal_palette.h
