+++ guacamole/src/protocols/ssh/terminal_scrollbar.h
