+++ guacamole/src/protocols/ssh/terminal_typescript.h
