+++ guacamole/src/protocols/ssh/terminal_xparsecolor.c
@@ -19,8 +19,8 @@
 
 #include "config.h"
 
-#include "terminal/named-colors.h"
-#include "terminal/palette.h"
+#include "terminal_named-colors.h"
+#include "terminal_palette.h"
 
 #include <stdio.h>
